Reverted the spliting up scripts
