+++ blocks/card-list/libraryCard.js
@@ -1,4 +1,4 @@
-import { makePublicUrl } from '../../scripts/url-helper.js';
+import { makePublicUrl } from '../../scripts/scripts.js';
 import { createOptimizedPicture } from '../../scripts/lib-franklin.js';
 import {
   li, a, p, div, h2,
